backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
docxtpl==0.20.1
python-docx==1.2.0
python-multipart==0.0.20
slowapi==0.1.9
google-generativeai==0.8.5
httpx==0.28.1
pydantic-settings==2.10.1
docx2pdf==0.1.8
